Elements: Add captions category
