+++ packages/docs/elements-sidebars.ts
@@ -26,6 +26,17 @@ const sidebars: SidebarsConfig = {
 		},
 		{
 			type: 'category',
+			label: 'Captions',
+			link: {type: 'doc', id: 'captions/index'},
+			collapsed: false,
+			items: [
+				'captions/moving-pill-captions/index',
+				'captions/popping-word-captions/index',
+				'captions/word-highlight-captions/index',
+			],
+		},
+		{
+			type: 'category',
 			label: 'Data',
 			link: {type: 'doc', id: 'data/index'},
 			collapsed: false,
@@ -56,9 +67,6 @@ const sidebars: SidebarsConfig = {
 				'text/news-article-headline-highlight/index',
 				'text/strike-through/index',
 				'text/text-marker/index',
-				'text/timed-captions-background/index',
-				'text/timed-captions-highlight/index',
-				'text/timed-captions-scale/index',
 			],
 		},
 	],
