+++ packages/docs/elements/captions/index.mdx
@@ -1,0 +1,10 @@
+---
+title: Captions
+description: Drop-in synchronized caption Elements for Remotion videos.
+---
+
+import {ElementLibrary} from '@site/src/components/Elements/ElementLibrary';
+
+# Captions
+
+<ElementLibrary category="captions" />
