+++ packages/docs/elements/captions/moving-pill-captions/index.mdx
@@ -1,0 +1,12 @@
+---
+title: Moving Pill Captions
+description: Synchronized captions with a pill that moves between spoken words.
+image: https://remotion.media/elements/captions-moving-pill-captions-preview.png
+---
+
+import {ElementPage} from '@site/src/components/Elements/ElementPage';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
+
+# Moving Pill Captions
+
+<ElementPage definition={elementDefinitions['captions/moving-pill-captions']} sourceFile="./moving-pill-captions.tsx" />
