+++ packages/docs/elements/captions/moving-pill-captions/moving-pill-captions.tsx
@@ -6,6 +6,7 @@ import React, {
 	forwardRef,
 	useEffect,
 	useImperativeHandle,
+	useLayoutEffect,
 	useMemo,
 	useRef,
 	useState,
@@ -25,15 +26,15 @@ import {
 	type SequenceProps,
 } from 'remotion';
 
-type TimedCaptionsScaleProps = InteractiveBaseProps &
+type MovingPillCaptionsProps = InteractiveBaseProps &
 	InteractiveTransformProps &
 	Pick<SequenceProps, 'width' | 'height'> & {
 		readonly captions?: Caption[];
 		readonly combineTokensWithinMilliseconds?: number;
 	};
 
-type TimedCaptionsScaleLayerProps = Omit<
-	TimedCaptionsScaleProps,
+type MovingPillCaptionsLayerProps = Omit<
+	MovingPillCaptionsProps,
 	'captions'
 > & {
 	readonly captions: Caption[];
@@ -43,11 +44,14 @@ const desiredFontSize = 80;
 const maximumTextWidth = 800;
 const fontWeight = '700';
 const textColor = '#ffffff';
-const highlightColor = '#18ff0e';
-const activeWordScale = 1.03;
+const backgroundColor = '#0b84f3';
+const pillHorizontalPadding = 12;
+const pillVerticalPadding = 12;
+const pillBorderRadius = 10;
+const pillMoveDurationInFrames = 5;
 const defaultCombineTokensWithinMilliseconds = 800;
 
-const timedCaptionsScaleSchema = {
+const movingPillCaptionsSchema = {
 	...Interactive.baseSchema,
 	...Interactive.captionsSchema,
 	width: {
@@ -103,47 +107,21 @@ const getActivePageIndex = (
 		),
 	);
 
-const getActiveTokenIndex = (
-	tokens: readonly Pick<TikTokToken, 'fromMs' | 'toMs'>[],
+const getLatestStartedTokenIndex = (
+	tokens: readonly Pick<TikTokToken, 'fromMs'>[],
 	timeMs: number,
 ) =>
-	tokens.findIndex((token) =>
-		isTimeWithinHalfOpenInterval(timeMs, token.fromMs, token.toMs),
+	tokens.reduce(
+		(latestIndex, token, tokenIndex) =>
+			token.fromMs <= timeMs ? tokenIndex : latestIndex,
+		-1,
 	);
 
-const getTokenScale = ({
-	currentTimeMs,
-	fps,
-	token,
-}: {
-	readonly currentTimeMs: number;
-	readonly fps: number;
-	readonly token: Pick<TikTokToken, 'fromMs' | 'toMs'>;
-}) => {
-	if (!isTimeWithinHalfOpenInterval(currentTimeMs, token.fromMs, token.toMs)) {
-		return 1;
-	}
-
-	const tokenDurationInFrames = ((token.toMs - token.fromMs) / 1000) * fps;
-	const tokenLocalFrame = ((currentTimeMs - token.fromMs) / 1000) * fps;
-	const animationDurationInFrames = Math.min(4, tokenDurationInFrames / 2);
-	const enterProgress = spring({
-		config: {damping: 200},
-		durationInFrames: animationDurationInFrames,
-		fps,
-		frame: tokenLocalFrame,
-	});
-	const exitProgress = interpolate(
-		tokenLocalFrame,
-		[tokenDurationInFrames - animationDurationInFrames, tokenDurationInFrames],
-		[1, 0],
-		{
-			extrapolateLeft: 'clamp',
-			extrapolateRight: 'clamp',
-		},
-	);
-
-	return 1 + Math.min(enterProgress, exitProgress) * (activeWordScale - 1);
+type TokenLayout = {
+	readonly height: number;
+	readonly left: number;
+	readonly top: number;
+	readonly width: number;
 };
 
 const CaptionPage: React.FC<{
@@ -158,7 +136,10 @@ const CaptionPage: React.FC<{
 			maximumTextWidth,
 			captionAreaWidth ?? maximumTextWidth,
 		);
-		const maximumTokenWidth = Math.max(1, availableWidth / activeWordScale);
+		const maximumTokenWidth = Math.max(
+			1,
+			availableWidth - pillHorizontalPadding * 2,
+		);
 		const tokenFontSizes = page.tokens
 			.map((token) => token.text.trim())
 			.filter(Boolean)
@@ -185,7 +166,98 @@ const CaptionPage: React.FC<{
 			...tokenFontSizes,
 		);
 	}, [captionAreaWidth, page.text, page.tokens]);
-	const activeTokenIndex = getActiveTokenIndex(page.tokens, currentTimeMs);
+	const textContainerRef = useRef<HTMLDivElement>(null);
+	const tokenRefs = useRef<Array<HTMLSpanElement | null>>([]);
+	const [tokenLayouts, setTokenLayouts] = useState<TokenLayout[]>([]);
+
+	useLayoutEffect(() => {
+		const container = textContainerRef.current;
+
+		if (!container) {
+			return;
+		}
+
+		const updateTokenLayouts = () => {
+			setTokenLayouts(
+				page.tokens.map((_, tokenIndex) => {
+					const token = tokenRefs.current[tokenIndex];
+
+					if (!token) {
+						return {height: 0, left: 0, top: 0, width: 0};
+					}
+
+					return {
+						height: token.offsetHeight,
+						left: token.offsetLeft,
+						top: token.offsetTop,
+						width: token.offsetWidth,
+					};
+				}),
+			);
+		};
+
+		updateTokenLayouts();
+		const resizeObserver = new ResizeObserver(updateTokenLayouts);
+		resizeObserver.observe(container);
+
+		return () => resizeObserver.disconnect();
+	}, [fontSize, page.tokens]);
+	const latestStartedTokenIndex = getLatestStartedTokenIndex(
+		page.tokens,
+		currentTimeMs,
+	);
+	const pageLocalFrame = ((currentTimeMs - page.startMs) / 1000) * fps;
+	const backgroundWordIndex = page.tokens.reduce(
+		(wordIndex, token, tokenIndex) => {
+			if (tokenIndex === 0) {
+				return wordIndex;
+			}
+
+			const tokenStartFrame = ((token.fromMs - page.startMs) / 1000) * fps;
+			return (
+				wordIndex +
+				spring({
+					config: {damping: 100},
+					delay: Math.max(0, tokenStartFrame - pillMoveDurationInFrames / 2),
+					durationInFrames: pillMoveDurationInFrames,
+					fps,
+					frame: pageLocalFrame,
+				})
+			);
+		},
+		0,
+	);
+	const hasTokenLayouts = tokenLayouts.length === page.tokens.length;
+	const clampedBackgroundWordIndex = Math.min(
+		Math.max(backgroundWordIndex, 0),
+		Math.max(0, tokenLayouts.length - 1),
+	);
+	const interpolationInput = tokenLayouts.map((_, tokenIndex) =>
+		Number(tokenIndex),
+	);
+	const interpolateTokenLayout = (values: number[]) => {
+		if (values.length === 0) {
+			return 0;
+		}
+
+		if (values.length === 1) {
+			return values[0];
+		}
+
+		return interpolate(clampedBackgroundWordIndex, interpolationInput, values);
+	};
+	const pillHeight = fontSize + pillVerticalPadding * 2;
+	const pillLeft = interpolateTokenLayout(
+		tokenLayouts.map((measurement) => measurement.left),
+	);
+	const pillTop = interpolateTokenLayout(
+		tokenLayouts.map(
+			(measurement) => measurement.top + (measurement.height - pillHeight) / 2,
+		),
+	);
+	const pillWidth = interpolateTokenLayout(
+		tokenLayouts.map((measurement) => measurement.width),
+	);
 	const textStrokeWidth = fontSize / 7;
 
 	return (
@@ -202,6 +274,7 @@ const CaptionPage: React.FC<{
 			}}
 		>
 			<div
+				ref={textContainerRef}
 				aria-hidden="true"
 				style={{
 					color: textColor,
@@ -211,14 +284,28 @@ const CaptionPage: React.FC<{
 					lineHeight: 1.5,
 					maxWidth: maximumTextWidth,
 					paintOrder: 'stroke fill',
+					position: 'relative',
 					textAlign: 'center',
 					WebkitTextStroke: `${textStrokeWidth}px #000000`,
 					whiteSpace: 'normal',
 					width: '100%',
 				}}
 			>
+				{hasTokenLayouts && latestStartedTokenIndex >= 0 ? (
+					<div
+						aria-hidden="true"
+						style={{
+							backgroundColor,
+							borderRadius: pillBorderRadius,
+							height: pillHeight,
+							left: pillLeft - pillHorizontalPadding,
+							position: 'absolute',
+							top: pillTop,
+							width: pillWidth + pillHorizontalPadding * 2,
+						}}
+					/>
+				) : null}
 				{page.tokens.map((token, tokenIndex) => {
-					const isActive = tokenIndex === activeTokenIndex;
 					const visibleText = token.text.trim();
 					const visibleTextIndex = token.text.indexOf(visibleText);
 					const leadingWhitespace = visibleText
@@ -234,12 +321,15 @@ const CaptionPage: React.FC<{
 						>
 							{leadingWhitespace}
 							<span
+								ref={(element) => {
+									tokenRefs.current[tokenIndex] = element;
+								}}
 								style={{
-									color: isActive ? highlightColor : textColor,
+									color: textColor,
 									display: 'inline-block',
-									scale: getTokenScale({currentTimeMs, fps, token}),
-									transformOrigin: 'center bottom',
+									position: 'relative',
 									whiteSpace: 'pre',
+									zIndex: 1,
 								}}
 							>
 								{visibleText}
@@ -253,7 +343,7 @@ const CaptionPage: React.FC<{
 	);
 };
 
-const TimedCaptionsScaleContent: React.FC<{
+const MovingPillCaptionsContent: React.FC<{
 	readonly captionAreaWidth: number | null;
 	readonly captions: Caption[];
 	readonly combineTokensWithinMilliseconds: number;
@@ -294,9 +384,9 @@ const TimedCaptionsScaleContent: React.FC<{
 	);
 };
 
-const TimedCaptionsScaleInner = forwardRef<
+const MovingPillCaptionsInner = forwardRef<
 	HTMLDivElement,
-	TimedCaptionsScaleLayerProps & {
+	MovingPillCaptionsLayerProps & {
 		readonly controls: SequenceControls | undefined;
 	}
 >(
@@ -335,7 +425,7 @@ const TimedCaptionsScaleInner = forwardRef<
 				layout="none"
 				{...interactiveProps}
 				controls={controls}
-				name={name ?? '<TimedCaptionsScale>'}
+				name={name ?? '<MovingPillCaptions>'}
 				outlineRef={outlineRef}
 			>
 				<div
@@ -346,7 +436,7 @@ const TimedCaptionsScaleInner = forwardRef<
 						...style,
 					}}
 				>
-					<TimedCaptionsScaleContent
+					<MovingPillCaptionsContent
 						captionAreaWidth={width ?? null}
 						captions={captions}
 						combineTokensWithinMilliseconds={combineTokensWithinMilliseconds}
@@ -358,20 +448,20 @@ const TimedCaptionsScaleInner = forwardRef<
 	},
 );
 
-const TimedCaptionsScaleLayer = Interactive.withSchema({
-	Component: TimedCaptionsScaleInner,
-	componentName: '<TimedCaptionsScale>',
+const MovingPillCaptionsLayer = Interactive.withSchema({
+	Component: MovingPillCaptionsInner,
+	componentName: '<MovingPillCaptions>',
 	componentIdentity: null,
-	schema: timedCaptionsScaleSchema,
+	schema: movingPillCaptionsSchema,
 	supportsEffects: false,
-}) as React.FC<TimedCaptionsScaleLayerProps>;
+}) as React.FC<MovingPillCaptionsLayerProps>;
 
-export const TimedCaptionsScale: React.FC<TimedCaptionsScaleProps> = ({
+export const MovingPillCaptions: React.FC<MovingPillCaptionsProps> = ({
 	captions,
 	...props
 }) => {
 	if (captions) {
-		return <TimedCaptionsScaleLayer {...props} captions={captions} />;
+		return <MovingPillCaptionsLayer {...props} captions={captions} />;
 	}
 
 	return (
@@ -384,7 +474,7 @@ export const TimedCaptionsScale: React.FC<TimedCaptionsScaleProps> = ({
 				width: 900,
 			}}
 		>
-			<TimedCaptionsScaleLayer
+			<MovingPillCaptionsLayer
 				{...props}
 				captions={[
 					{
