+++ packages/docs/elements/captions/popping-word-captions/index.mdx
@@ -1,0 +1,12 @@
+---
+title: Popping Word Captions
+description: Synchronized captions that pop each spoken word into focus.
+image: https://remotion.media/elements/captions-popping-word-captions-preview.png
+---
+
+import {ElementPage} from '@site/src/components/Elements/ElementPage';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
+
+# Popping Word Captions
+
+<ElementPage definition={elementDefinitions['captions/popping-word-captions']} sourceFile="./popping-word-captions.tsx" />
