+++ packages/docs/elements/captions/word-highlight-captions/index.mdx
@@ -1,0 +1,12 @@
+---
+title: Word Highlight Captions
+description: Synchronized captions that highlight each spoken word.
+image: https://remotion.media/elements/captions-word-highlight-captions-preview.png
+---
+
+import {ElementPage} from '@site/src/components/Elements/ElementPage';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
+
+# Word Highlight Captions
+
+<ElementPage definition={elementDefinitions['captions/word-highlight-captions']} sourceFile="./word-highlight-captions.tsx" />
