--- packages/docs/elements/text/timed-captions-background/index.mdx
@@ -1,12 +1,0 @@
----
-title: Background Timed Captions
-description: Animated captions with word-level moving background styling.
-image: https://remotion.media/elements/text-timed-captions-background-preview.png
----
-
-import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
-
-# Background Timed Captions
-
-<ElementPage definition={elementDefinitions['text/timed-captions-background']} sourceFile="./timed-captions-background.tsx" />
