--- packages/docs/elements/text/timed-captions-highlight/index.mdx
@@ -1,12 +1,0 @@
----
-title: Highlight Timed Captions
-description: Animated captions with word-level highlight styling.
-image: https://remotion.media/elements/text-timed-captions-highlight-preview.png
----
-
-import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
-
-# Highlight Timed Captions
-
-<ElementPage definition={elementDefinitions['text/timed-captions-highlight']} sourceFile="./timed-captions-highlight.tsx" />
