--- packages/docs/elements/text/timed-captions-scale/index.mdx
@@ -1,12 +1,0 @@
----
-title: Scale Timed Captions
-description: Animated captions with word-level scale styling.
-image: https://remotion.media/elements/text-timed-captions-scale-preview.png
----
-
-import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
-
-# Scale Timed Captions
-
-<ElementPage definition={elementDefinitions['text/timed-captions-scale']} sourceFile="./timed-captions-scale.tsx" />
